Create SimonSays.js
This code creates a basic Simon Says game that you can play in your web browser. The game randomly generates a sequence of colors and waits for the player to repeat the sequence by clicking on the colored buttons. If the player gets the sequence right, the game adds another color to the pattern and continues. If the player makes a mistake, the game ends.